+++ lib/src/gg_multi_version.dart
@@ -9,4 +9,4 @@
 // coverage:ignore-file
 
 /// The version of the `gg_multi` package.
-const String ggMultiVersion = '7.12.0';
+const String ggMultiVersion = '7.13.0';