+++ test/backend/organization_test.dart
@@ -4,8 +4,8 @@
 // Use of this source code is governed by terms that can be
 // found in the LICENSE file in the root of this package.
 
-import 'package:test/test.dart';
 import 'package:gg_multi/src/backend/organization.dart';
+import 'package:test/test.dart';
 
 void main() {
   group('Organization model', () {