+++ test/backend/vscode_launcher_test.dart
@@ -5,8 +5,9 @@
 // found in the LICENSE file in the root of this package.
 
 import 'dart:io';
-import 'package:test/test.dart';
+
 import 'package:gg_multi/src/backend/vscode_launcher.dart';
+import 'package:test/test.dart';
 
 void main() {
   group('VSCodeLauncher', () {