+++ test/commands/do/ls_test.dart
@@ -10,9 +10,9 @@ import 'package:args/command_runner.dart';
 import 'package:gg_args/gg_args.dart';
 import 'package:gg_capture_print/gg_capture_print.dart';
 import 'package:gg_multi/src/backend/constants.dart';
+import 'package:gg_multi/src/commands/do/ls.dart';
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
-import 'package:gg_multi/src/commands/do/ls.dart';
 
 void main() {
   group('ListCommand', () {