+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 7.13.0 - 2026-08-04
+
+### Changed
+
+- Show ticket deletion dialog at the beginning of publish. Recolorize.
+- Use overrides in package.json
+
 ## 7.12.0 - 2026-08-04
 
 ## 7.11.2 - 2026-08-04